PostMapping lesson

+++ become-spring-boot-dev/spring-REST/tennis-player-rest/src/main/java/io/datajek/tennisplayerrest/PlayerRepository.java
@@ -1,6 +1,7 @@
 package io.datajek.tennisplayerrest;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlayerRepository extends JpaRepository<Player, Integer> {
+public interface PlayerRepository extends JpaRepository<Player, Long> {
 
 }
